% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener.R
\name{assert}
\alias{assert}
\title{Terminates the execution with an error message if the given conditional statement is \code{FALSE}}
\usage{
assert(flag, err_msg = "Assertion Error !", err_src = NULL)
}
\arguments{
\item{flag}{A boolean: Condition to be checked}

\item{err_msg}{A character string: Message to be displayed}
}
\description{
Terminates the execution with an error message if the given conditional statement is \code{FALSE}
}
\examples{
x = 1
assert(x > 1, "Argument x can not be greater than 1")

Error in assert(x > 1, "Argument x can not be greater than 1") :
  Argument x can not be greater than 1

}
