% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener.R
\name{datetimes}
\alias{datetimes}
\title{Returns labels of the columns of the given data frame containing date-time values}
\usage{
datetimes(df)
}
\value{
vector of characters containing date-time column labels of the given data.frame
}
\description{
Returns labels of the columns of the given data frame containing date-time values
}
\section{Fields}{

\describe{
\item{\code{df}}{data.frame: The table for which date-time columns are required}
}}

