% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener.R
\name{feasible}
\alias{feasible}
\title{Returns if a given value meets a given feasibility condition. The feasibility condition is specified by
a feasibility range and a set of non-feasible values.}
\usage{
feasible(v, range = NULL, excludes = NULL)
}
\arguments{
\item{v}{A numeric or a vector of numerics or integers}

\item{range}{Vector of two numerics containing the lower and upper bounds of the feasibility range.
If \code{range} is NULL (default), range condition checking is skipped.}

\item{exclude}{Vector of numerics containing non-feasible values
If \code{exclude} is NULL (default), non-feasibility condition checking is skipped.}
}
\value{
Logical TRUE if the given value is not missing, within the given range and not in \code{exclude},
                FALSE otherwise.
}
\description{
Returns if a given value meets a given feasibility condition. The feasibility condition is specified by
a feasibility range and a set of non-feasible values.
}
