% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener.R
\name{high.pass.mean}
\alias{high.pass.mean}
\title{Returns mean of a vector where values lower than a given threshold are filtered}
\usage{
high.pass.mean(v, threshold = 0, na.rm = T)
}
\arguments{
\item{v}{numeric or integer: Given vector for which the filtered mean is computed}

\item{threshold}{numeric (scalar): value specifying the filtering threshold (default is \code{0})}

\item{na.rm}{logical: Should missing values be filtered? (default is TRUE)}
}
\value{
numeric (scalar): Filtered mean computed
}
\description{
Returns mean of a vector where values lower than a given threshold are filtered
}
\examples{
high.pass.mean(c(1, NA, 3, 2, -1, -2), threshold = 1)
[1] 2.5

}
