% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener.R
\name{is.empty}
\alias{is.empty}
\title{Use this function to check if a variable is empty or not}
\usage{
is.empty(x)
}
\arguments{
\item{x}{A variable. Can be a vector,matrix,list or data.frame}
}
\value{
A boolean: TRUE if argument \code{x} is empty, FALSE otherwise
}
\description{
Use this function to check if a variable is empty or not
}
\examples{
a = matrix()
is.empty(a)
[1] TRUE
a = c(2, 4)
is.empty(a)
[1] FALSE

}
