% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener.R
\name{make.err.msg}
\alias{make.err.msg}
\title{Makes a pretty error message to be displayed on the console}
\usage{
make.err.msg(err_msg = "", err_src = "gener::make.err.msg")
}
\arguments{
\item{err_msg}{A character string: Message to be displayed}

\item{err_src}{A character string: Source from which the error message is issued.}
}
\value{
A character string containing the modified error message
}
\description{
Makes a pretty error message to be displayed on the console
}
\examples{
stop(make.err.msg(err_msg = "Some error message", err_src = match.call()[[1]]))

Error:

Error from: match.call:
------------------------

Some error message

}
