% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener.R
\name{most.common}
\alias{most.common}
\title{Returns the most frequent element of given vector \code{v}}
\usage{
most.common(v)
}
\arguments{
\item{v}{Vector}
}
\value{
The most frequent element of given vector \code{v}
}
\description{
Returns the most frequent element of given vector \code{v}
}
