% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener.R
\name{nameColumns}
\alias{nameColumns}
\title{This function, changes the column names of the given table (data.frame) and checks for the column classes}
\usage{
nameColumns(dataset, columns, classes = NULL, convert = T)
}
\arguments{
\item{dataset}{A table(data.frame) on which the column names is going to change}

\item{columns}{A named list of characters indicating which names should change to what.
List values should be a subset of original column names and list names are names to which the original column names should change}

\item{classes}{a named list of characters inducating the required classes of each column.
List names should be the new column names and values are classes required for each column}

\item{convert}{A boolean indicating if columns should be coerced to the required class. Function returns error is coercing is not possible.
Defaulted is TRUE. If FALSE, the function verifies the column classes and returns error if don't match.}
}
\value{
A table(data.frame) same class as the input table \code{dataset}
}
\description{
This function, changes the column names of the given table (data.frame) and checks for the column classes
}
\examples{
Data = data.frame(x = c(1,2,3), y = c('10', '20', '30'), z = c('One', 'two', 'three'), stringsAsFactors = F)
Data \%>\% nameColumns(columns = list(XX = 'x', YY = 'y', ZZ = 'z'), classes = list(XX = 'integer', YY = 'numeric'))

}
