% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener.R
\name{no.missing}
\alias{no.missing}
\title{returns TRUE if none of the elements of given argument v is missing}
\usage{
no.missing(v)
}
\description{
returns TRUE if none of the elements of given argument v is missing
}
