% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener.R
\name{nominals}
\alias{nominals}
\title{Returns nominal(categorical) column labels of a data frame}
\usage{
nominals(df)
}
\value{
vector of characters containing nominal column labels of the given table
}
\description{
Returns nominal(categorical) column labels of a data frame
}
\section{Fields}{

\describe{
\item{\code{df}}{data.frame: The table for which nominal columns are required}
}}

