% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener.R
\name{numerics}
\alias{numerics}
\title{Returns numeric column labels of a data frame}
\usage{
numerics(df)
}
\value{
vector of characters containing numeric column labels of the given table
}
\description{
Returns numeric column labels of a data frame
}
\section{Fields}{

\describe{
\item{\code{df}}{data.frame: The table for which numeric columns are required}
}}

