% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener.R
\name{repeat.char}
\alias{repeat.char}
\title{Replicates the given charachter to make a longer character string}
\usage{
repeat.char(char, N)
}
\arguments{
\item{char}{a character: Character to be replicated}

\item{N}{An integer: number of times \code{char} should be repeated}
}
\value{
A character string
}
\description{
Replicates the given charachter to make a longer character string
}
\examples{
repeat.char('X', 3)
[1] "XXX"

}
