% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener.R
\name{roundto.multiple}
\alias{roundto.multiple}
\title{Rounds \code{x} to the nearest multiple of \code{N}}
\usage{
roundto.multiple(x, N, adjust = "closest")
}
\arguments{
\item{x}{A numeric or integer}

\item{N}{A numeric or integer}

\item{adjust}{A character string:
'closest': Returns the closest multiple of \code{N} \cr
'top'   ': Returns the closest multiple of \code{N} that is greater than \code{x} \cr
'bottom' : Returns the closest multiple of \code{N} that is smaller than \code{x} \cr}
}
\value{
A numeric
}
\description{
Rounds \code{x} to the nearest multiple of \code{N}
}
\examples{
roundto.multiple(345.45, 1)
[1] 345
roundto.multiple(345.45, 7)
[1] 343
roundto.multiple(345.45, 7.2)/7.2
[1] 48

}
