% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener.R
\name{toVectorList}
\alias{toVectorList}
\title{Converts a data.frame or matrix into a list of named vectors where each column is a named vector}
\usage{
toVectorList(df)
}
\description{
Converts a data.frame or matrix into a list of named vectors where each column is a named vector
}
