% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener.R
\name{vect.extend}
\alias{vect.extend}
\title{Extends the given vector \code{v} to length \code{N}}
\usage{
vect.extend(v, N)
}
\arguments{
\item{v}{Vector of any type}

\item{N}{An integer}
}
\value{
Vector same type as \code{v}
}
\description{
Extends the given vector \code{v} to length \code{N}
}
\examples{
a = c(1, 2, 3)
vect.extend(a, 5)
[1] 1 2 3 1 2

}
