% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener.R
\name{vect.map}
\alias{vect.map}
\title{Maps a given vector into a given range}
\usage{
vect.map(x, minimum = 0, maximum = 1)
}
\arguments{
\item{x}{A numeric vector: Vector to be mapped}

\item{minimum}{A numeric scalar: Specifies the lower bound of the output}

\item{maximum}{A numeric scalar: Specifies the higher bound of the output}
}
\value{
A numeric vector
}
\description{
Maps a given vector into a given range
}
\examples{
vect.map(c(1, 2, 3), minimum = 100, maximum = 300)
[1] 100 200 300

}
