% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{vect.normalize}
\alias{vect.normalize}
\title{Normalizes the given vector by dividing each element by the vector magnitude}
\usage{
vect.normalize(v, cent = FALSE)
}
\arguments{
\item{v}{a vector of numerics to be normalized}

\item{cent}{A logical. Should the vector be also centralized?}
}
\value{
A numeric vector containing the normalized values. Sum of squares of elements of the output vector will be 1.0.
}
\description{
Normalizes the given vector by dividing each element by the vector magnitude
}
\examples{
vect.normalize(c(3, 4))
[1] 0.6 0.8

}
